fix: users migration foreign key

+++ wallet-api/migrations/20220406010733_create_users_table.ts
@@ -4,7 +4,7 @@ export async function up(knex: Knex): Promise<void> {
   return knex.schema.createTableIfNotExists("users", builder => {
     builder.increments("id")
     builder.string("cpf", 11).notNullable().unique()
-    builder.bigInteger("account_id").notNullable()
+    builder.integer("account_id").notNullable()
     builder.foreign("account_id").references("id").inTable("accounts")
     builder.timestamps(true, true)
   })
